+++ packages/number-input/src/Arrows/Arrows.types.ts
@@ -1,0 +1,13 @@
+import { Direction } from '../NumberInput/NumberInput.types';
+
+export interface ArrowProps {
+  /**
+   * Determines if buttons should be disabled
+   */
+  disabled: boolean;
+
+  /**
+   * Callback called on button interaction
+   */
+  handleValueChange: (value: Direction) => void;
+}
