+++ packages/number-input/src/Arrows/index.tsx
@@ -1,0 +1,1 @@
+export { Arrows } from './Arrows';
