+++ packages/number-input/src/Input/Input.tsx
@@ -2,20 +2,14 @@ import React, { ChangeEvent } from 'react';
 
 import { cx } from '@leafygreen-ui/emotion';
 import { useControlledValue, useForwardedRef } from '@leafygreen-ui/hooks';
-import Icon from '@leafygreen-ui/icon';
 import WarningIcon from '@leafygreen-ui/icon/dist/Warning';
 import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
 import { createSyntheticEvent } from '@leafygreen-ui/lib';
 
-import { Size, State } from '../NumberInput/NumberInput.types';
+import { Arrows } from '../Arrows';
+import { Direction, Size, State } from '../NumberInput/NumberInput.types';
 
 import {
-  arrowBaseStyles,
-  arrowDisabledStyles,
-  arrowsAnimateStyles,
-  arrowsBaseStyles,
-  arrowThemeStyles,
-  downArrowRotateStyles,
   iconBaseStyles,
   iconDisabledStyles,
   iconErrorDisabledStyles,
@@ -46,7 +40,7 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
     {
       value: valueProp,
       onChange: onChangeProp,
-      disabled,
+      disabled = false,
       size = Size.Default,
       hasSelectOptions,
       state = State.None,
@@ -72,40 +66,18 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
       handleChange?.(synthEvent as ChangeEvent<HTMLInputElement>);
     };
 
-    const handleValueChange = (value: 'increment' | 'decrement') => {
+    const handleValueChange = (value: Direction) => {
       switch (value) {
-        case 'increment': {
+        case Direction.Increment: {
           inputRef.current?.stepUp();
           handleSyntheticEvent();
           break;
         }
 
-        case 'decrement': {
+        case Direction.Decrement: {
           inputRef.current?.stepDown();
           handleSyntheticEvent();
           break;
-        }
-      }
-    };
-
-    /**
-     * Edge case if the user clicks on an arrow button then switches to keyboard click.
-     * By default if focus is in the input then the keyboard clicks will work automatically but since the buttons are custom and outside of the input we are mimicking native behavior.
-     */
-    const handleArrowKeyDown = (e: React.KeyboardEvent) => {
-      if (!disabled) {
-        switch (e.key) {
-          case 'ArrowUp': {
-            e.preventDefault();
-            handleValueChange('increment');
-            break;
-          }
-
-          case 'ArrowDown': {
-            e.preventDefault();
-            handleValueChange('decrement');
-            break;
-          }
         }
       }
     };
@@ -160,39 +132,7 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
           <WarningIcon aria-hidden="true" />
         </div>
 
-        <div
-          className={cx(arrowsBaseStyles, {
-            [arrowsAnimateStyles]: !disabled,
-            [arrowDisabledStyles]: disabled,
-          })}
-        >
-          <button
-            aria-label="Increment number"
-            onClick={() => handleValueChange('increment')}
-            onKeyDown={handleArrowKeyDown}
-            className={cx(arrowBaseStyles, arrowThemeStyles[theme])}
-            type="button"
-            tabIndex={-1} // Mimicking native behavior; you cannot focus on an arrow.
-          >
-            <Icon aria-hidden={true} glyph="CaretUp" size={16} />
-          </button>
-          <button
-            aria-label="Decrement number"
-            onClick={() => handleValueChange('decrement')}
-            onKeyDown={handleArrowKeyDown}
-            className={cx(arrowBaseStyles, arrowThemeStyles[theme])}
-            type="button"
-            tabIndex={-1}
-          >
-            <Icon
-              className={downArrowRotateStyles}
-              aria-hidden={true}
-              // TODO: CaretUp and CaretDown icons do not align vertically
-              glyph="CaretUp"
-              size={16}
-            />
-          </button>
-        </div>
+        <Arrows disabled={disabled} handleValueChange={handleValueChange} />
       </div>
     );
   },
